Move case model processor to j2me code where it is used
